+++ plone/app/drafts/lifecycle.py
@@ -2,7 +2,7 @@
 from plone.app.drafts.interfaces import ICurrentDraftManagement
 from plone.app.drafts.interfaces import IDraftStorage
 from plone.app.drafts.utils import getCurrentDraft
-from plone.app.drafts.utils import getObjectKey
+from plone.app.drafts.utils import getDefaultKey
 from plone.app.drafts.utils import syncDraft
 from zope.component import queryUtility
 
@@ -21,14 +21,14 @@ def beginDrafting(context, event):
     if storage is None or not storage.enabled:
         return
 
-    request = getattr(context, 'REQUEST', None)
+    request = getattr(context, "REQUEST", None)
     if request is None:
         return
 
     current = ICurrentDraftManagement(request)
 
     # Update target key regardless - we could have a stale cookie
-    current.targetKey = getObjectKey(context)
+    current.targetKey = getDefaultKey(context)
 
     if current.draftName is None:
         drafts = storage.getDrafts(current.userId, current.targetKey)
@@ -52,7 +52,7 @@ def syncDraftOnSave(context, event):
     if storage is None or not storage.enabled:
         return
 
-    request = getattr(context, 'REQUEST', None)
+    request = getattr(context, "REQUEST", None)
     if request is None:
         return
 
@@ -76,7 +76,7 @@ def discardDraftsOnCancel(context, event):
     if storage is None or not storage.enabled:
         return
 
-    request = getattr(context, 'REQUEST', None)
+    request = getattr(context, "REQUEST", None)
     if request is None:
         return