+++ plone/app/drafts/utils.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 from AccessControl import getSecurityManager
-from Acquisition import aq_parent
 from plone.app.drafts.interfaces import ICurrentDraftManagement
 from plone.app.drafts.interfaces import IDraftStorage
 from plone.app.drafts.interfaces import IDraftSyncer
 from plone.uuid.interfaces import IUUID
-from Products.CMFCore.utils import getToolByName
 from zope.component import getAdapters
 from zope.component import queryUtility
 
@@ -58,35 +56,11 @@ def getCurrentDraft(request, create=False):
 
 
 def getCurrentUserId():
-    """Get the current user id. Returns None if the user is Anonymous.
-    """
+    """Get the current user id. Returns None if the user is Anonymous."""
 
     return getSecurityManager().getUser().getId()
 
 
 def getDefaultKey(context):
-    """Get the default (string) key for the given context, based on uuids
-    """
+    """Get the default (string) key for the given context, based on uuids"""
     return IUUID(context, None)
-
-
-def getObjectKey(context):
-    """Get a key for an Archetypes object. This will be a string
-    representation of its uuid, unless it is in the portal_factory, in
-    which case it'll be the a string like
-    "${parent_uuid}:portal_factory/${portal_type}"
-    """
-
-    portal_factory = getToolByName(context, 'portal_factory', None)
-    if portal_factory is None or not portal_factory.isTemporary(context):
-        return getDefaultKey(context)
-
-    tempFolder = aq_parent(context)
-    folder = aq_parent(aq_parent(tempFolder))
-
-    defaultKey = getDefaultKey(folder)
-    if defaultKey is None:
-        # probably the portal root
-        defaultKey = '0'
-
-    return '{0}:{1}'.format(defaultKey, tempFolder.getId())